% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cohortwise_cov.R
\name{calculate_cohortwise_cov}
\alias{calculate_cohortwise_cov}
\title{calculate_cohortwise_cov}
\usage{
calculate_cohortwise_cov(raw_matrix, metadata, cohort_col = "Cohort")
}
\arguments{
\item{raw_matrix}{matrix/dataframe containing raw values.}

\item{metadata}{dataframe containing metadata information}

\item{cohort_col}{A metadata column where cohorts are present}
}
\value{
A dataframe with calculated cov
}
\description{
calcualte coefficient of variation across all cohorts
}
\examples{
calculate_cohortwise_cov(raw_matrix, metadata, cohort_col='Cohort')
}
