% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_differential_expression.R
\name{compute_differential_expression}
\alias{compute_differential_expression}
\title{compute_differential_expression}
\usage{
compute_differential_expression(
  norm_data = NULL,
  metadata = NULL,
  cohort_col = NULL,
  cohort_a = NULL,
  cohort_b = NULL,
  algo = "limma",
  paired = FALSE,
  equal_var = TRUE,
  nonpar = FALSE,
  pval_adjust_method = "BH",
  fc_with_raw = FALSE,
  raw_data = NULL,
  raw_log_flag = FALSE,
  expr_stat = TRUE
)
}
\arguments{
\item{norm_data}{The dataframe/matrix with samples in columns and features in rows.}

\item{metadata}{The dataframe containing metadata information}

\item{cohort_col}{A metadata column where cohorts are present}

\item{cohort_a}{Vector of cohorts used as cohort_a}

\item{cohort_b}{Vector of cohorts used as cohort_b}

\item{algo}{Use limma or ttest}

\item{paired}{A logical indicating whether to use paired samples for ttest or fold change caculations.}

\item{equal_var}{A logical variable indicating whether to treat the two variances as being equal. 
If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used. 
It is only applicable with ttest algorithm.}

\item{nonpar}{A logical variable indicating whether to perform a non-parametric test i.e. wilcox.test (TRUE) or t.test. It is only applicable with ttest algorithm.}

\item{pval_adjust_method}{Provide pval adjust method ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"). Also check p.adjust from stats.}

\item{fc_with_raw}{Use raw data (data without normalization) to calculate fold change.}

\item{raw_data}{The dataframe/matrix with samples in columns and features in rows having same dimensions as norm data.}

\item{raw_log_flag}{Check if raw data is log transformed (TRUE) or not (FALSE).}

\item{expr_stat}{Calculate MaxExpr and MinExpr (TRUE) where MaxExpr/MinExpr is calculated as the maximum/minimum of average of samples within cohorts. 
If fc_with_raw = TRUE then it will use raw data from expression statistics calculation.}
}
\value{
The differential expression result with state1 vs state2 (state1/state2) or cohort-B vs cohort-A (cohort-B/cohort-A) comparison, If logFC >0, it implies abundance is greater in cohort-B (state1).
}
\description{
It calculates the differential expression for cohort-B vs cohort-A or cohort-B/cohort-A comparison
}
\examples{
compute_differential_expression(norm_data, metadata, 'Cohort', 'Cohort1', 'Cohort2')
}
