% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fold_change.R
\name{compute_fold_change}
\alias{compute_fold_change}
\title{compute_fold_change}
\usage{
compute_fold_change(
  data_mat = NULL,
  samples_a = NULL,
  samples_b = NULL,
  paired = FALSE,
  log_flag = FALSE
)
}
\arguments{
\item{data_mat}{A dataframe/matrix containing samples in columns and features in rows.}

\item{samples_a}{A vector of samples in cohort A.}

\item{samples_b}{A vector of samples in cohort B.}

\item{paired}{A logical indicating whether to use paired samples for fold change calculation.}

\item{log_flag}{A logical variable indicating whether data is log transformed or not. 
If TRUE then fold change will be calculated as B - A (logFC) otherwise B/A (FC).}
}
\value{
A dataframe with calculated statistics.
}
\description{
Calculate fold change on dataframe/matrix where samples are in columns and features are in rows.
}
\examples{
compute_fold_change(data_mat, samples_a, samples_b)
}
