% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_opls_distance.R
\name{compute_opls_distance}
\alias{compute_opls_distance}
\title{compute_opls_distance}
\usage{
compute_opls_distance(opl, significance = TRUE, level = 0.05)
}
\arguments{
\item{opl}{The "opls" object (ropls library) with typeC = 'OPLS-DA'.}

\item{significance}{A boolean indicating whether significant points are to be computed.}

\item{level}{A float indicating the significance level to be used in the two-sided chi-square test. Ignored is signiicance=FALSE.}
}
\value{
If significance=TRUE, A list of 1) dataframe containg scores distance data, 
2) a vector containing distances corresponding to significance level, and, 
3) A vector of indices of significant points in the dataframe.
Otherwise only (1) is returned.
}
\description{
Computes the component distance data required for outlier diagnistics.
}
