% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pca.R
\name{compute_pca}
\alias{compute_pca}
\title{compute_pca}
\usage{
compute_pca(sample_raw_mat)
}
\arguments{
\item{sample_raw_mat}{matrix/dataframe containing raw values.}
}
\value{
A list of loadings and pc values
}
\description{
performs pca on the sample raw matrix
}
