% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_s_data.R
\name{compute_s_data}
\alias{compute_s_data}
\title{compute_s_data}
\usage{
compute_s_data(opl = NULL)
}
\arguments{
\item{opl}{An opls object}
}
\value{
A dataframe
}
\description{
Computes the loading and vorrelation dataframe required for S-plot.
}
\examples{
compute_s_data(opls_object)
}
