% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_t_test.R
\name{compute_t_test}
\alias{compute_t_test}
\title{compute_t_test}
\usage{
compute_t_test(
  data_mat = NULL,
  samples_a = NULL,
  samples_b = NULL,
  paired = FALSE,
  equal_var = TRUE,
  nonpar = FALSE,
  pval_adjust_method = "BH"
)
}
\arguments{
\item{data_mat}{A dataframe/matrix containing samples in columns and features in rows.}

\item{samples_a}{A vector of samples in cohort A.}

\item{samples_b}{A vector of samples in cohort B.}

\item{paired}{A logical indicating whether to perform paired t-test or not.}

\item{equal_var}{A logical variable indicating whether to treat the two variances as being equal. 
If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{nonpar}{A logical variable indicating whether to perform a non-parametric test i.e. wilcox.test (TRUE) or t.test.}
}
\value{
A dataframe with calculated statistics.
}
\description{
Perform t-test on dataframe/matrix where samples are in columns and features are in rows.
}
\examples{
compute_t_test(data_mat, samples_a, samples_b)
}
