% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_boxplot_on_matrix.R
\name{create_boxplot_on_matrix}
\alias{create_boxplot_on_matrix}
\title{create_boxplot_on_matrix}
\usage{
create_boxplot_on_matrix(
  sample_raw_mat = NULL,
  x_label = "",
  y_label = "",
  title_label = "",
  flip_coord = FALSE,
  plot_axis_format = "plain",
  plot_axis_text_size = 10
)
}
\arguments{
\item{sample_raw_mat}{sample_raw_mat matrix/dataframe containing raw values}

\item{x_label}{Label x-axis}

\item{y_label}{Label y-axis}

\item{title_label}{Title of the plot}

\item{flip_coord}{Flip Coordinates}

\item{plot_axis_format}{font face ("plain", "italic", "bold", "bold.italic")}

\item{plot_axis_text_size}{set axis text size}
}
\value{
ggplot object
}
\description{
Makes boxplot on matrix
}
\examples{
create_boxplot_on_matrix(sample_raw_mat = NULL, x_label = "Sample", y_label = "Raw Intensity", plot_axis_format = 'plain', plot_axis_text_size = 10)
}
