% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohortwise_cov_barplot.R
\name{create_cohortwise_cov_barplot}
\alias{create_cohortwise_cov_barplot}
\title{create_cohortwise_cov_barplot}
\usage{
create_cohortwise_cov_barplot(
  calculated_cov_df,
  id_order = NULL,
  id_col = "id",
  cohorts_order = NULL
)
}
\arguments{
\item{calculated_cov_df}{A dataframe with calculated cov}

\item{cohorts_order}{The order of cohorts}
}
\value{
A plotly object
}
\description{
Makes boxplot for cov values across cohorts
}
\examples{
create_cohortwise_cov_barplot(calculated_cov_df, id_order = c('cmpd1','cmpd2'), id_col = 'id', cohorts_order = c('CohortA','CohortB'))
}
