% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohortwise_cov_boxplot.R
\name{create_cohortwise_cov_boxplot}
\alias{create_cohortwise_cov_boxplot}
\title{create_cohortwise_cov_boxplot}
\usage{
create_cohortwise_cov_boxplot(calculated_cov_df, cohorts_order = NULL)
}
\arguments{
\item{calculated_cov_df}{A dataframe with calculated cov}

\item{cohorts_order}{The order of cohorts}
}
\value{
A plotly object
}
\description{
Makes boxplot for cov values across cohorts
}
\examples{
create_cohortwise_cov_boxplot(calculated_cov_df, cohorts_order = c('CohortA','CohortB'))
}
