% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohortwise_identifier_barplot.R
\name{create_cohortwise_identifier_barplot}
\alias{create_cohortwise_identifier_barplot}
\title{create_cohortwise_identifier_barplot}
\usage{
create_cohortwise_identifier_barplot(
  sample_raw_mat = NULL,
  metadata_df = NULL,
  id_order = NULL,
  cohorts_order = NULL,
  cohort_col = "Cohort",
  interactive = FALSE,
  x_label = "",
  y_label = "",
  title_label = ""
)
}
\arguments{
\item{sample_raw_mat}{sample_raw_mat matrix/dataframe containing raw values}

\item{metadata_df}{A dataframe with samples to cohort mapping}

\item{id_order}{A vector of id's for which barplot to be made}

\item{cohorts_order}{The order of cohorts}

\item{cohort_col}{A cohort column present in metadata}

\item{interactive}{Make plot interactive using plotly}

\item{x_label}{Label x-axis}

\item{y_label}{Label y-axis}

\item{title_label}{Title of the plot}
}
\value{
ggplot object or plotly object
}
\description{
Makes cohortwise barplot for selected identifiers
}
\examples{
create_cohortwise_identifier_barplot(sample_raw_mat = NULL, metadata_df = NULL, id_order = NULL, 
                          cohorts_order = NULL, cohort_col='Cohort', 
                          x_label = "", y_label = "", title_label = "")
}
