% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multi_cohorts_metadata_for_anova.R
\name{create_multi_cohorts_metadata_for_anova}
\alias{create_multi_cohorts_metadata_for_anova}
\title{create_multi_cohorts_metadata_for_anova}
\usage{
create_multi_cohorts_metadata_for_anova(metadata = NULL, cohort_col = NULL)
}
\arguments{
\item{metadata}{The dataframe containing metadata information}

\item{cohort_col}{A vector of multiple columns from metadata used to make combinations of columns}
}
\value{
dataframe, metadata with newly added columns
}
\description{
It adds the different combinations of columns after pasting them as new columns in metadata.
}
\examples{
create_multi_cohorts_metadata_for_anova(metadata, cohort_col = c("Cohort", "Time"))
}
