% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multi_cohorts_metadata_for_limma.R
\name{create_multi_cohorts_metadata_for_limma}
\alias{create_multi_cohorts_metadata_for_limma}
\title{create_multi_cohorts_metadata_for_limma}
\usage{
create_multi_cohorts_metadata_for_limma(
  metadata = NULL,
  cohort_col = NULL,
  cohort_a = NULL,
  cohort_b = NULL,
  new_colname = NULL
)
}
\arguments{
\item{metadata}{The dataframe containing metadata information}

\item{cohort_col}{A metadata column where cohorts are present}

\item{cohort_a}{Vector of cohorts used as cohort_a}

\item{cohort_b}{Vector of cohorts used as cohort_b}

\item{new_colname}{New column where new conditions will be placed}
}
\value{
dataframe, metadata with newly added column
}
\description{
It creates a metadata of two new conditions by pasting cohorts within cohort_a and cohort_b.
}
\examples{
create_multi_cohorts_metadata_for_limma(metadata, 'Cohort', c("Cohort1", "Cohort2"),  c("Cohort3", "Cohort4"), "multi_condition")
}
