% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_opls.R
\name{create_opls}
\alias{create_opls}
\title{create_opls}
\usage{
create_opls(
  sample_data,
  metadata,
  condition,
  scalC = c("none", "center", "pareto", "standard")[4]
)
}
\arguments{
\item{sample_data}{Input dataframe with shape (n x m) where n is the number of samples
and m = number of variables}

\item{metadata}{Sample metadata dataframe with number of rows = n}

\item{condition}{A column in metadata on which the OPLS-DA has to be performed. Should
have binary classes.}

\item{scalC}{Character: either no centering nor scaling ('none'), mean-centering only ('center'),
mean-centering and pareto scaling ('pareto'), or mean-centering and unit variance scaling ('standard') (default)}
}
\value{
An 'opls' class object (see library 'ropls')
}
\description{
Creates a class opls object for OPLS-DA
}
\examples{
 
create_opls(sample_data, metadata, condiition)
}
