% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_opls_score_data.R
\name{create_opls_score_data}
\alias{create_opls_score_data}
\title{create_opls_score_data}
\usage{
create_opls_score_data(opl)
}
\arguments{
\item{opl}{The "opls" object (ropls library) with typeC = 'OPLS-DA' for which scores are required.}
}
\value{
A dataframe of scores and orthogonal component.
}
\description{
Function to create the scores data dataframe of OPLS-DA.
}
