% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pairwise_posthoc_boxplot.R
\name{create_pairwise_posthoc_boxplot}
\alias{create_pairwise_posthoc_boxplot}
\title{create_pairwise_posthoc_boxplot}
\usage{
create_pairwise_posthoc_boxplot(
  posthoc_data_long_df,
  intensity_data,
  selected_metabolite,
  selected_comparison
)
}
\arguments{
\item{posthoc_data_long_df}{A dataframe containing long format post-hoc data.}

\item{intensity_data}{A dataframe containing intensity data.}

\item{selected_metabolite}{The selected metabolite for analysis.}

\item{selected_comparison}{The selected comparison for analysis.}
}
\value{
A plotly object representing the pairwise post-hoc boxplot.
}
\description{
Creates a pairwise post-hoc boxplot for selected metabolite and comparison.
}
