% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_supertest_sets.R
\name{create_supertest_sets}
\alias{create_supertest_sets}
\title{create_supertest_sets}
\usage{
create_supertest_sets(sets_list)
}
\arguments{
\item{sets_list}{A list of sets to be compared.}
}
\description{
It provides the sets of common ids shared between different sets in upset plot.
}
\examples{
create_supertest_sets(sets_list)
}
