% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_upset_plot.R
\name{create_upset_plot}
\alias{create_upset_plot}
\title{create_upset_plot}
\usage{
create_upset_plot(
  sets_list,
  nsets = 20,
  nintersects = 100,
  sets = NULL,
  keep.order = F,
  set.metadata = NULL,
  intersections = NULL,
  matrix.color = "gray23",
  main.bar.color = "gray23",
  mainbar.y.label = "Intersection Size",
  mainbar.y.max = NULL,
  sets.bar.color = "gray23",
  sets.x.label = "Set Size",
  point.size = 3,
  line.size = 1.8,
  mb.ratio = c(0.7, 0.3),
  expression = NULL,
  att.pos = NULL,
  att.color = main.bar.color,
  order.by = c("freq", "degree"),
  decreasing = c(T, F),
  show.numbers = "yes",
  number.angles = 0,
  group.by = "degree",
  cutoff = NULL,
  queries = NULL,
  query.legend = "none",
  shade.color = "gray88",
  shade.alpha = 0.25,
  matrix.dot.alpha = 0.5,
  empty.intersections = NULL,
  color.pal = 1,
  boxplot.summary = NULL,
  attribute.plots = NULL,
  scale.intersections = "identity",
  scale.sets = "identity",
  text.scale = 1.1,
  set_size.angles = 0,
  set_size.show = TRUE,
  set_size.numbers_size = NULL,
  set_size.scale_max = NULL
)
}
\arguments{
\item{sets_list}{A list of sets to be compared.}

\item{nsets}{Number of sets to look at}

\item{nintersects}{Number of intersections to plot. If set to NA, all intersections will be plotted.}

\item{sets}{Specific sets to look at (Include as combinations. Ex: c("Name1", "Name2"))}
}
\description{
It creates an UpSet plot from list of vectors elements. For more information type help(UpSetR::upset)
}
\examples{
create_upset_plot(sets_list, nsets = 5, nintersects = 10, sets = NULL)
}
