% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp_limma.R
\name{diff_exp_limma}
\alias{diff_exp_limma}
\title{diff_exp_limma}
\usage{
diff_exp_limma(
  sample_raw_mat = NULL,
  metadata = NULL,
  cohort_col = NULL,
  cohort_a = NULL,
  cohort_b = NULL,
  pval_adjust_method = "BH",
  log_flag = TRUE
)
}
\arguments{
\item{sample_raw_mat}{The sample matrix with samples in columns and ids as rownames.}

\item{metadata}{The dataframe containing metadata information}

\item{cohort_col}{A metadata column where cohorts are present}

\item{cohort_a}{Vector of cohorts used as cohort_a}

\item{cohort_b}{Vector of cohorts used as cohort_b}

\item{pval_adjust_method}{Provide pval adjust method ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"). Also check p.adjust from stats.}

\item{log_flag}{Check if data is log transformed (TRUE) or not (FALSE). If not (FALSE) then do internally log2 transformation.}
}
\value{
dataframe, If logFC >0, it implies abundance is greater in cohort_b, MaxExpr is calculated as the maximum of average of samples within cohorts.
}
\description{
calls limma differential expression algorithms
}
\examples{
diff_exp_limma(sample_raw_mat, metadata, 'Cohort', 'Cohort1', 'Cohort2')
}
