% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_matrix_by_intensity.R
\name{filter_matrix_by_intensity}
\alias{filter_matrix_by_intensity}
\title{filter_matrix_by_intensity}
\usage{
filter_matrix_by_intensity(
  sample_raw_mat = NULL,
  algo = "sample",
  metadata_df = NULL,
  cohort_col = NULL,
  intensity_threshold = NULL,
  samples_threshold = NULL
)
}
\arguments{
\item{sample_raw_mat}{A dataframe containing samples with raw values}

\item{algo}{An algo used to filter data by all samples (algo = "sample") or at least in one cohort (algo = "cohort")}

\item{metadata_df}{A dataframe with samples to cohort mapping}

\item{cohort_col}{A cohort column present in metadata}

\item{intensity_threshold}{The minimum intensity that should be present in samples}

\item{samples_threshold}{The number of samples in percentage where minimum intensity should be present}
}
\value{
The filtered sample_raw_mat
}
\description{
Filter matrix by intensity present in at least x (in percent) samples across all samples or in at least one cohort.
}
\examples{
filter_matrix_by_intensity(sample_raw_mat, algo = "sample")
}
