% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_metadata_by_cohorts.R
\name{filter_metadata_by_cohorts}
\alias{filter_metadata_by_cohorts}
\title{filter_metadata_by_cohorts}
\usage{
filter_metadata_by_cohorts(
  metadata,
  condition = "Cohort",
  selected_cohorts = NULL
)
}
\arguments{
\item{condition}{A metadata column where cohorts are present.}

\item{selected_cohorts}{A vector of cohorts}

\item{metadata_df}{A dataframe with samples to cohort mapping.}
}
\value{
A dataframe of filtered metadata.
}
\description{
performs pca on the sample raw matrix
}
\examples{
filter_metadata_by_cohorts(metadata_df, 'Cohort', c("Cohort1, Cohort2"))
}
