% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_data.R
\name{impute_missing_data}
\alias{impute_missing_data}
\title{impute_missing_data}
\usage{
impute_missing_data(sample_raw_mat = NULL, method = NULL)
}
\arguments{
\item{sample_raw_mat}{A dataframe/matrix containing samples in columns and features in rows}

\item{method}{The method used to impute missing data}
}
\value{
The dataframe with imputated missing data
}
\description{
Impute missing data using the following methods: 
"exclude": Exclude features having missing values; 
"replace_by_zero": Replace missing values by zero; 
"feature_lod": Replace missing values by limit of detection (LoD, 1/5 of the minimum positive value of each feature); 
"feature_min": Replace missing values by 1/2 of the minimum values of each feature row; 
"feature_mean": Replace missing values by the mean values of each feature row; 
"feature_median": Replace missing values by the median values of each feature row; 
"feature_knn": Estimate missing values by k-nearest neighbours based on similar features - KNN (feature-wise); 
"sample_knn": Estimate missing values by k-nearest neighbours based on similar samples - KNN (sample-wise); 
"bpca": Estimate missing values by Bayesian PCA (BPCA) method; 
"ppca": Estimate missing values by probabilistic PCA (PPCA) method; 
"svdImpute": Estimate missing values by Singular Value Decomposition (SVD) method;
}
\examples{
impute_missing_data(sample_raw_mat, method)
}
