% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_intomix_input.R
\name{make_intomix_input}
\alias{make_intomix_input}
\title{make_intomix_input}
\usage{
make_intomix_input(
  norm_data = NULL,
  metadata = NULL,
  cohorts = NULL,
  cohorts_compare_data = NULL,
  cohort_col = "Cohort",
  rownames_col = "uniqueId",
  remove_duplicates = TRUE,
  id_col = "compoundId",
  p_val_correct_methods = "BH",
  log_flag = TRUE
)
}
\arguments{
\item{norm_data}{A sample intenisty data with samples in columns.}

\item{metadata}{dataframe containing metadata information}

\item{cohorts}{A vector of cohorts used for cohorts comparisons}

\item{cohorts_compare_data}{A dataframe having predefined cohorts comparisons}

\item{cohort_col}{A metadata column where cohorts are present}

\item{rownames_col}{A raw_intensity_df column which is to be used for assigning rownames.}

\item{remove_duplicates}{Remove duplicate ids rows based on pval}

\item{id_col}{A column id which will be used to remove duplicates}

\item{p_val_correct_methods}{character vector containing p-val correction methods ('Bonferroni', 'BH(FDR)') to be applied.}

\item{log_flag}{Check if data is log transformed (TRUE) or not (FALSE). If not (FALSE) then do internally log2 transformation.}
}
\value{
A dataframe of differential expression for all cohorts comparison.
}
\description{
Create an intomix input file
}
\examples{
make_intomix_input(norm_data = NULL, metadata = NULL, cohorts = NULL,  
                   cohorts_compare_data = NULL, cohort_col = 'Cohort', 
                   rownames_col = 'uniqueId', remove_duplicates = TRUE, 
                   id_col = 'compoundId', p_val_correct_methods = 'BH',
                   log_flag = TRUE )
}
