% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_log2fc_matrix_from_intomix_input.R
\name{make_log2fc_matrix_from_intomix_input}
\alias{make_log2fc_matrix_from_intomix_input}
\title{make_log2fc_matrix_from_intomix_input}
\usage{
make_log2fc_matrix_from_intomix_input(
  intomix_input = NULL,
  selected_comparisons = NULL,
  pval_cutoff = 0.05
)
}
\arguments{
\item{intomix_input}{The intomix input}

\item{selected_comparisons}{The vector of selected comparisons in "state1 vs state2" format}

\item{pval_cutoff}{The pval cutoff to filter the data}
}
\value{
A dataframe of comparisons in columns and ids in rows
}
\description{
It creates the log2FC matrix from the intomix input format
}
\examples{
make_log2fc_matrix_from_intomix_input(intomix_input, selected_comparisons, pval_cutoff = 0.05)
}
