% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_by_scaling_factor.R
\name{normalize_by_scaling_factor}
\alias{normalize_by_scaling_factor}
\title{normalize_by_scaling_factor}
\usage{
normalize_by_scaling_factor(
  sample_raw_mat,
  normalization_agent,
  scaling_factor_col = 1
)
}
\arguments{
\item{sample_raw_mat}{A dataframe/matrix containing samples with raw values.}

\item{normalization_agent}{A dataframe/matrix with samples as rownames and having numeric columns}

\item{scaling_factor_col}{A numeric column name which is present in normalization_agent dataframe}
}
\value{
A dataframe/marix with normlazed data
}
\description{
normaize data by samplewise scaling factor
}
\examples{
normalize_by_scaling_factor(sample_raw_mat, normalization_agent, scaling_factor_col = 1)
}
