% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_way_anova_on_matrix.R
\name{one_way_anova_on_matrix}
\alias{one_way_anova_on_matrix}
\title{one_way_anova_on_matrix}
\usage{
one_way_anova_on_matrix(
  sample_raw_mat = NULL,
  metadata_df = NULL,
  cohort_col = "Cohort"
)
}
\arguments{
\item{sample_raw_mat}{A dataframe containing samples with raw values.}

\item{metadata_df}{A dataframe with samples to cohort mapping.}

\item{cohort_col}{A metadata column where cohorts are present.}
}
\value{
A dataframe with two extra columns (calculated F.Value and P.Value).
}
\description{
performs one way anova test on the sample raw matrix
}
\examples{
one_way_anova_on_matrix(sample_raw_mat, metadata_df, cohort_col = "Cohort")
}
