% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opls_s_plot.R
\name{opls_s_plot}
\alias{opls_s_plot}
\title{opls_s_plot}
\usage{
opls_s_plot(
  opl,
  x_label = "Loadings",
  y_label = "Correlation",
  title_label = "S-Plot for OPLS-DA"
)
}
\arguments{
\item{opl}{The computed "opls" object (ropls library) on which the plot has to be made.}

\item{x_label}{The label for x axis.}

\item{y_label}{The label for y axis.}

\item{title_label}{The title for the plot}
}
\value{
A plotly plot
}
\description{
opls_s_plot
}
