% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opls_scores_data.R
\name{opls_scores_data}
\alias{opls_scores_data}
\title{opls_scores_data}
\usage{
opls_scores_data(opl)
}
\arguments{
\item{opl}{The "opls" object (ropls library) with typeC = 'OPLS-DA' for which scores are required.}
}
\value{
A dataframe of scores and orthogonal component.
}
\description{
opls_scores_data
}
\examples{
data(scaurine)
attach(sacurine)
w = opls(dataMatrix,sampleMetadata[,'gender'],predI=1,orthoI=2)
opls_scores_data(w)

}
