% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opls_scores_plot.R
\name{opls_scores_plot}
\alias{opls_scores_plot}
\title{opls_scores_plot}
\usage{
opls_scores_plot(
  scores_data,
  metadata = NULL,
  condition = NULL,
  show_ellipse = FALSE,
  interactive = TRUE,
  title_label = "",
  marker_size = 6,
  title_label_size = 18,
  axis_title_size = 14,
  opls_cohort_text_format = "bold",
  opls_cohort_text_align = "right",
  opls_cohort_title_size = 18,
  opls_cohort_sample_size = 15,
  opls_plot_axis_format = "bold",
  opls_plot_axis_text_size = 14
)
}
\arguments{
\item{scores_data}{A dataframe containing the scores data.}

\item{metadata}{Dataframe containing samples to cohort mapping information. The first column must be ssame as the rownames in in the scores_data.}

\item{condition}{A metadata column where cohorts are present}

\item{interactive}{make plot interactive (default is TRUE)}

\item{title_label}{Title of the plot}

\item{marker_size}{The size of marker point}

\item{title_label_size}{Size of title label}

\item{axis_title_size}{Size of axis title}

\item{opls_cohort_text_format}{set text format of cohort legends}

\item{opls_cohort_text_align}{align cohort legends}

\item{opls_cohort_title_size}{set font size of cohort title}

\item{opls_cohort_sample_size}{set font size of cohorts}

\item{opls_plot_axis_format}{set axis format}

\item{opls_plot_axis_text_size}{set axis text size}

\item{significance_data}{A numeric of length 2 containing the significance level distances of both axes.}
}
\value{
A plotly object
}
\description{
Plots the model scores vs orthogonal score for each sample.
}
\examples{
opls_scores_plot(scores_data,metadat)

}
