% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_normalization.R
\name{perform_normalization}
\alias{perform_normalization}
\title{perform_normalization}
\usage{
perform_normalization(sample_raw_mat = NULL, norm_type = NULL)
}
\arguments{
\item{sample_raw_mat}{matrix/dataframe containing raw values}

\item{norm_type}{Select normalization type from the following methods:
"auto_scale": mean-centered and divided by the standard deviation of each variable;
"pareto_scale": mean-centered and divided by the square root of the standard deviation of each variable;
"mean_center_scale": mean-centered only;  
"range_scale": mean-centered and divided by the range of each variable;
"log_trans": log transformation;
"log10_trans": log10 transformation;
"log2_trans": log2 transformation;
"glog_trans": generalized logarithm transformation, tolerant to 0 and negative values;
"glog10_trans": generalized logarithm transformation with base 10, tolerant to 0 and negative values;
"glog2_trans": generalized logarithm transformation with base 2, tolerant to 0 and negative values;
"asinh_trans": ArcSinh transformation;}
}
\value{
The normalized sample_raw_mat
}
\description{
It normalizes the data using some generic methods
}
\examples{
perform_normalization(sample_raw_mat, norm_type)
}
