% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_opls_s3d.R
\name{plot_opls_s3d}
\alias{plot_opls_s3d}
\title{plot_opls_s3d}
\usage{
plot_opls_s3d(
  s_data = NULL,
  x_label = "Loadings",
  y_label = "Correlation",
  title_label = "S-Plot for OPLS-DA"
)
}
\arguments{
\item{s_data}{A dataframe of size (n x 2) with the two columns corresponding to the Loadings and Correlation.}

\item{x_label}{The label for x axis.}

\item{y_label}{The label for y axis.}

\item{title_label}{The title for the plot}
}
\value{
A plotly plot
}
\description{
Plots the S-plot for OPLS-DA loadings.
}
