% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca3d.R
\name{plot_pca3d}
\alias{plot_pca3d}
\title{plot_pca3d}
\usage{
plot_pca3d(
  PCAObj_Summary = NULL,
  metadata = NULL,
  condition = NULL,
  color_palette = NULL,
  pc_x = 1,
  pc_y = 2,
  pc_z = 3,
  title_label = "",
  show_legend_title = TRUE
)
}
\arguments{
\item{PCAObj_Summary}{A list of summary of prcomp function.}

\item{metadata}{dataframe containing samples to cohort mapping information}

\item{condition}{A metadata column where cohorts are present}

\item{color_palette}{The named vector with colors as values and cohorts as keys}

\item{pc_x}{PC to keep on x-axis}

\item{pc_y}{PC to keep on y-axis}

\item{pc_z}{PC to keep on z-axis}

\item{title_label}{Title of the plot}

\item{show_legend_title}{Show legend title (TRUE/FALSE)}
}
\value{
plotly object
}
\description{
makes a plotly PCA-3D plot
}
\examples{
plot_pca3d(PCAObj_Summary, metadata, 'Cohort', pc_x = 1, pc_y = 2, pc_z= 3)
}
