% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_proportion_of_variance.R
\name{plot_proportion_of_variance}
\alias{plot_proportion_of_variance}
\title{plot_proportion_of_variance}
\usage{
plot_proportion_of_variance(PCAObj_Summary)
}
\arguments{
\item{PCAObj_Summary}{A list of summary of prcomp function.}
}
\value{
A plotly object
}
\description{
performs pca on the sample raw matrix
}
\examples{
plot_proportion_of_variance(PCAObj_Summary)
}
