% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_features_with_constant_data.R
\name{remove_features_with_constant_data}
\alias{remove_features_with_constant_data}
\title{remove_features_with_constant_data}
\usage{
remove_features_with_constant_data(sample_raw_mat = NULL)
}
\arguments{
\item{sample_raw_mat}{A dataframe/matrix containing samples in columns and features in rows}
}
\value{
The filtered sample_raw_mat
}
\description{
Remove features having contant values in all samples
}
\examples{
remove_features_with_constant_data(sample_raw_mat)
}
