% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_features_with_missing_data.R
\name{remove_features_with_missing_data}
\alias{remove_features_with_missing_data}
\title{remove_features_with_missing_data}
\usage{
remove_features_with_missing_data(sample_raw_mat = NULL, samples_cutoff = 50)
}
\arguments{
\item{sample_raw_mat}{A dataframe/matrix containing samples in columns and features in rows}

\item{samples_cutoff}{The cutoff of the number of samples in percentage where missing values are allowed}
}
\value{
The filtered sample_raw_mat
}
\description{
Remove features having missing values in samples above the specified cutoff
}
\examples{
remove_features_with_missing_data(sample_raw_mat, samples_cutoff)
}
