% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_intensity_matrix.R
\name{sample_intensity_matrix}
\alias{sample_intensity_matrix}
\title{sample_intensity_matrix}
\usage{
sample_intensity_matrix(
  raw_intensity_df = NULL,
  metadata_df = NULL,
  rownames_col = NULL,
  drop_na = FALSE,
  replace_na_with_zero = TRUE
)
}
\arguments{
\item{raw_intensity_df}{A dataframe containing samples with raw values.}

\item{metadata_df}{A dataframe with samples to cohort mapping.}

\item{rownames_col}{A raw_intensity_df column which is to be used for assigning rownames.}

\item{drop_na}{Drop rows having NA's.}

\item{replace_na_with_zero}{Replace all NA's with zero.}
}
\value{
A marix with samples in columns and ids in rows with raw values.
}
\description{
create a sample matrix with samples in columns and ids in rownames.
}
\examples{
sample_intensity_matrix(raw_intensity_df, metadata_df, rownames_col = 'Id')
}
