% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_string_every_nth_char.R
\name{split_string_every_nth_char}
\alias{split_string_every_nth_char}
\title{split_string_every_nth_char}
\usage{
split_string_every_nth_char(string = NULL, n = 20)
}
\arguments{
\item{string}{A character string or vector of strings}

\item{n}{The number to split string after every nth position}
}
\value{
A vector of splitted string
}
\description{
This function splits a string after every nth position
}
\examples{
split_string_every_nth_char(string)
}
