% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{volcano_plot}
\usage{
volcano_plot(
  diff_exp_rdesc = NULL,
  log2fc_range = NULL,
  p_val_cutoff = NULL,
  fdr_cutoff = NULL
)
}
\arguments{
\item{diff_exp_rdesc}{The differential expression dataframe from limma}

\item{log2fc_range}{Vector of log2FC range}

\item{p_val_cutoff}{The pval cutoff}

\item{fdr_cutoff}{The FDR cutoff}
}
\value{
plotly object
}
\description{
makes a plotly volcano plot
}
