% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_anova.R
\name{compute_anova}
\alias{compute_anova}
\title{compute_anova}
\usage{
compute_anova(sample_raw_mat = NULL, metadata_df = NULL, cohort_col = "Cohort")
}
\arguments{
\item{sample_raw_mat}{A dataframe containing samples with raw values.}

\item{metadata_df}{A dataframe with samples to cohort mapping.}

\item{cohort_col}{A vector of metadata columns used for n way anova.}
}
\value{
A dataframe with calculated F.Value and P.Value.
}
\description{
performs n way anova test on the sample raw matrix
}
\examples{
compute_anova(sample_raw_mat, metadata_df, cohort_col = "Cohort")
}
