% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohortwise_cov_barplot.R
\name{create_cohortwise_cov_barplot}
\alias{create_cohortwise_cov_barplot}
\title{create_cohortwise_cov_barplot}
\usage{
create_cohortwise_cov_barplot(
  calculated_cov_df,
  id_order = NULL,
  id_col = "id",
  cohorts_order = NULL,
  show_legend = TRUE,
  interactive = FALSE
)
}
\arguments{
\item{calculated_cov_df}{A dataframe with calculated cov}

\item{id_order}{A vector of ids to plot}

\item{id_col}{The column name where ids are present}

\item{cohorts_order}{The order of cohorts}

\item{show_legend}{Show Lagend on plot}

\item{interactive}{make plot interactive if set TRUE}
}
\value{
A plotly object
}
\description{
Makes boxplot for cov values across cohorts
}
\examples{
create_cohortwise_cov_barplot(calculated_cov_df, id_order = c('cmpd1','cmpd2'), id_col = 'id', cohorts_order = c('CohortA','CohortB'))
}
