% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cohortwise_identifier_boxplot.R
\name{create_cohortwise_identifier_boxplot}
\alias{create_cohortwise_identifier_boxplot}
\title{create_cohortwise_identifier_boxplot}
\usage{
create_cohortwise_identifier_boxplot(
  sample_raw_mat = NULL,
  metadata_df = NULL,
  id_order = NULL,
  cohorts_order = NULL,
  cohort_col = "Cohort",
  x_label = NULL,
  y_label = NULL,
  title_label = NULL,
  interactive = FALSE
)
}
\arguments{
\item{sample_raw_mat}{sample_raw_mat matrix/dataframe containing raw values}

\item{metadata_df}{A dataframe with samples to cohort mapping}

\item{id_order}{A vector of id's for which barplot to be made}

\item{cohorts_order}{The order of cohorts}

\item{cohort_col}{A cohort column present in metadata}

\item{x_label}{Label x-axis}

\item{y_label}{Label y-axis}

\item{title_label}{Title of the plot}

\item{interactive}{Make plot interactive using plotly}
}
\value{
ggplot object or plotly object
}
\description{
Makes cohortwise boxplot for selected identifiers
}
\examples{
create_cohortwise_identifier_boxplot(sample_raw_mat, metadata_df, id_order, cohorts_order, cohort_col)
}
