% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_densityplot_on_matrix.R
\name{create_densityplot_on_matrix}
\alias{create_densityplot_on_matrix}
\title{create_densityplot_on_matrix}
\usage{
create_densityplot_on_matrix(
  sample_raw_mat = NULL,
  data_type = "all",
  col_name = NULL
)
}
\arguments{
\item{sample_raw_mat}{sample_raw_mat matrix/dataframe containing raw values}

\item{data_type}{use datafrom whole matrix or a column ("all" or "column")}

\item{col_name}{A column name from sample_raw_mat}
}
\value{
ggplot object
}
\description{
Makes densityplot on matrix
}
\examples{
create_densityplot_on_matrix(sample_raw_mat = NULL, data_type = "all", col_name = NULL)
}
