% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_exp_limma.R
\name{diff_exp_limma}
\alias{diff_exp_limma}
\title{diff_exp_limma}
\usage{
diff_exp_limma(
  prot_norm_mat,
  metadata,
  cohort_condition,
  cohort_a,
  cohort_b,
  p_val_correct_methods,
  log_flag
)
}
\arguments{
\item{prot_norm_mat}{matrix/dataframe containing protein normalised abundances.}

\item{metadata}{dataframe containing metadata information}

\item{cohort_condition}{A metadata column where cohorts are present}

\item{cohort_a}{string cohort_a}

\item{cohort_b}{string cohort_b}

\item{p_val_correct_methods}{character vector containing p-val correction methods ('Bonferroni', 'BH(FDR)') to be applied.}

\item{log_flag}{log normalises the data and performes limma.}
}
\value{
dataframe, If logFC >0, it implies abundance is greater in cohort_b.
}
\description{
calls limma differential expression algorithms
}
