% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_matrix_by_most_variance.R
\name{filter_matrix_by_most_variance}
\alias{filter_matrix_by_most_variance}
\title{filter_matrix_by_most_variance}
\usage{
filter_matrix_by_most_variance(
  sample_raw_mat = NULL,
  top_n = NULL,
  algo = "mad",
  metadata_df = NULL,
  cohort_col = NULL,
  remove_duplicates = TRUE,
  row_desc = NULL,
  id_col = NULL
)
}
\arguments{
\item{sample_raw_mat}{A dataframe containing samples with raw values.}

\item{top_n}{Filter top n number of ids'.}

\item{algo}{An algorithm used to sort ids', options are anova and mad.}

\item{metadata_df}{A dataframe with samples to cohort mapping.}

\item{cohort_col}{A metadata column where cohorts are present.}

\item{remove_duplicates}{remove duplicates using some column id from row descriptors.}

\item{row_desc}{A dataframe of row descriptors of the matrix.}

\item{id_col}{The id column of row descriptors dataframe used to remove duplicates.}
}
\value{
The filtered sample_raw_mat.
}
\description{
Filter matrix by most variance
}
\examples{
filter_matrix_by_most_variance(sample_raw_mat, top_n = NULL, algo = "anova", metadata_df, cohort_col = "Cohort")
}
