% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{plot_pca}
\usage{
plot_pca(pca_loadings_var_list, metadata, condition, pc_x, pc_y)
}
\arguments{
\item{pca_loadings_var_list}{list containing the PC loadings for every sample and the variances explained, returned by the compute_pca function.
This returns a plotly object of the PCA plot for all the samples.}

\item{metadata}{dataframe containing samples to cohort mapping information}

\item{condition}{A metadata column where cohorts are present}

\item{pc_x}{PC to keep on x-axis}

\item{pc_y}{PC to keep on y-axis}
}
\value{
plotly object
}
\description{
makes a plotly PCA plot
}
