% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{plot_pca}
\usage{
plot_pca(
  PCAObj_Summary,
  metadata,
  condition,
  pc_x = 1,
  pc_y = 2,
  show_ellipse = FALSE,
  interactive = TRUE,
  color_palette = NULL,
  title_label = "",
  marker_size = 6,
  title_label_size = 18,
  axis_title_size = 14,
  pca_cohort_text_format = "bold",
  pca_cohort_text_align = "right",
  pca_cohort_title_size = 18,
  pca_cohort_sample_size = 15,
  pca_plot_axis_format = "bold",
  pca_plot_axis_text_size = 14
)
}
\arguments{
\item{PCAObj_Summary}{A list of summary of prcomp function.}

\item{metadata}{dataframe containing samples to cohort mapping information}

\item{condition}{A metadata column where cohorts are present}

\item{pc_x}{PC to keep on x-axis}

\item{pc_y}{PC to keep on y-axis}

\item{show_ellipse}{show ellipse on plot (default is FALSE)}

\item{interactive}{make plot interactive (default is TRUE)}

\item{color_palette}{The vector of colors}

\item{title_label}{Title of the plot}

\item{marker_size}{The size of marker point}

\item{title_label_size}{Size of title label}

\item{axis_title_size}{Size of axis title}

\item{pca_cohort_text_format}{set text format of cohort legends}

\item{pca_cohort_text_align}{align cohort legends}

\item{pca_cohort_title_size}{set font size of cohort title}

\item{pca_cohort_sample_size}{set font size of cohorts}

\item{pca_plot_axis_format}{set axis format}

\item{pca_plot_axis_text_size}{set axis text size}
}
\value{
plotly object
}
\description{
makes a plotly PCA plot
}
\examples{
plot_pca(PCAObj_Summary, metadata, 'Cohort', pc_x = 1, pc_y = 2, interactive = TRUE, pca_cohort_text_format= 'bold', pca_cohort_text_align= "right",
         pca_cohort_title_size= 18, pca_cohort_sample_size= 15, pca_plot_axis_format= 'bold', pca_plot_axis_text_size= 14)
}
