% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca3d.R
\name{plot_pca3d}
\alias{plot_pca3d}
\title{plot_pca3d}
\usage{
plot_pca3d(PCAObj_Summary, metadata, condition, pc_x = 1, pc_y = 2, pc_z = 3)
}
\arguments{
\item{PCAObj_Summary}{A list of summary of prcomp function.}

\item{metadata}{dataframe containing samples to cohort mapping information}

\item{condition}{A metadata column where cohorts are present}

\item{pc_x}{PC to keep on x-axis}

\item{pc_y}{PC to keep on y-axis}

\item{pc_z}{PC to keep on z-axis}
}
\value{
plotly object
}
\description{
makes a plotly PCA plot
}
\examples{
plot_pca3d(PCAObj_Summary, metadata, 'Cohort', pc_x = 1, pc_y = 2, pc_z= 3)
}
