% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano_from_limma.R
\name{plot_volcano_from_limma}
\alias{plot_volcano_from_limma}
\title{volcano_plot}
\usage{
plot_volcano_from_limma(
  diff_exp_rdesc = NULL,
  log2fc_range = NULL,
  p_val_cutoff = NULL,
  fdr_cutoff = NULL,
  annotate_genes = NULL,
  interactive = TRUE
)
}
\arguments{
\item{diff_exp_rdesc}{The differential expression dataframe from limma}

\item{log2fc_range}{Vector of log2FC range}

\item{p_val_cutoff}{The pval cutoff}

\item{fdr_cutoff}{The FDR cutoff}

\item{annotate_genes}{A vector of genes/metabolites to be annotated}

\item{interactive}{make plot interactive (default is TRUE)}
}
\value{
plotly object
}
\description{
makes a plotly volcano plot
}
\examples{
plot_volcano_from_limma(diff_exp_rdesc, log2fc_range = 0.5, p_val_cutoff = 0.05, fdr_cutoff = NULL,
 annotate_genes = c('a','b'), interactive = TRUE)
}
